requirements.txt
# Flask==2.1.2
# SQLAlchemy==1.4.41
# PyMySQL==1.0.2
# gunicorn==20.1.0
Flask==2.2.5
pymysql==1.1.0
Flask-Bcrypt==1.0.1
gunicorn==23.0.0
python-dotenv==1.0.1